requirements.txt
aiogram
aiohttp
requests
python-dotenv
SQLALCHEMY
APScHEDULER
aiosqlite
gitHub
pandas
matplotlib



